+++ apps/shop/src/app/(shop)/my/orders/page.tsx
@@ -1,12 +1,6 @@
 import { OrderHistoryList, fetchOrderHistory } from '@/features/order';
 import { getCookieString } from '@/shared/utils';
 
-export default async function MyOrdersPage() {
-  const orderHistory = await getOrderHistory();
-
-  return <OrderHistoryList history={orderHistory} />;
-}
-
 async function getOrderHistory(page = 1) {
   const cookieString = await getCookieString();
   return await fetchOrderHistory(page, {
@@ -14,4 +8,10 @@ async function getOrderHistory(page = 1) {
       Cookie: cookieString,
     },
   });
+}
+
+export default async function MyOrdersPage() {
+  const orderHistory = await getOrderHistory();
+
+  return <OrderHistoryList history={orderHistory} />;
 }
