+++ apps/shop/src/features/order/components/OrderHistoryList.module.scss
@@ -1,7 +1,7 @@
 .article {
   @include flex-col;
   margin-top: 16px;
-  min-height: 60vh;
+  min-height: 80vh;
 }
 
 .pagination {
